optional header unittest and bugfixes

+++ src/main/java/com/github/katjahahn/StandardEntry.java
@@ -10,7 +10,7 @@ public class StandardEntry {
 
 	public String key;
 	public String description;
-	public int value;
+	public long value;
 
 	/**
 	 * @constructor Creates a standard entry with the values specified.
@@ -19,7 +19,7 @@ public class StandardEntry {
 	 * @param description
 	 * @param value
 	 */
-	public StandardEntry(String key, String description, int value) {
+	public StandardEntry(String key, String description, long value) {
 		this.key = key;
 		this.description = description;
 		this.value = value;
@@ -27,7 +27,7 @@ public class StandardEntry {
 
 	@Override
 	public String toString() {
-		return description + ": " + value + " (0x" + Integer.toHexString(value)
+		return description + ": " + value + " (0x" + Long.toHexString(value)
 				+ ")";
 	}
 }
